Performing persistence operations with repository

+++ src/main/java/io/github/kleitonmarques/libraryapi/model/Livro.java
@@ -1,9 +1,7 @@
 package io.github.kleitonmarques.libraryapi.model;
 
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
@@ -33,7 +31,7 @@ public class Livro {
     private GeneroLivro genero;
 
     @Column(name = "preco", precision = 18, scale = 2)
-    private Double preco;
+    private BigDecimal preco;
 
     @ManyToOne
     @JoinColumn(name = "id_autor")
